Change tests for link processing to be only string based
Before, the tests where creating minimal html containing a link that needs to be modified.
This results in the implementations to iterate over the whole document multiple
times to replace all the links.

Now the separate link processing functions should only operate on the href string
value.

Once this is inplace, then an additional function should be created to iterate over
all links and pass the necessary values to the fixer functions.