Add breakpoints for at least medium screens
Then the screen is wider than 550px, the navigation is shown horizontally.
If the screen is also taller than 500px the collapsability is disabled.

This is derived from iPhone screen dimensions. When the phone is vertical,
then the menu can be collapsed and the nagivation is displayed vertically.
When the phone is held horizontally (which is wide but not tall), then
the menu is horizonal, but can still be collapsed.

Only on taller screen is the menu always visible (and also horizontal).