Add font swap and use Georgia as fallback
I notices in production that the loading time for the custom Baskerville font takes
quite a moment. To still make the text availavle in that period I have now set
`font-display: swap`. While the custom font is loading, this displays the content
with the next front in the font stack.

Once the custom font is fully loaded, the broswers that support the font-display option
will swap out the displayed font for the custom font. To avoid jumping of the content
during this transistion as much as possible, I have set the second font in the
font stack to Gerorgia. It is a bit slimmer than Bakserville, but the letter height is comparable.
This should reduce the jumping during font swap better than using Times as the fallback.